Revert "alloc int8 so there is no conversion from 4 to 8, see calc_zhveed_size" Changing to Int8 introduces problems with zheevds liwork parameter. This reverts commit f8847b3ea3cab5248456ab8044b6ae6fa9b1e0e9.

+++ src/helper.f90
@@ -69,7 +69,7 @@ contains
       character(len=1)          :: vn_flag
       complex(8), intent(in)    :: H(:,:)
       real(8), intent(in)       :: eig_val(:)
-      integer(8), intent(out)   :: lwork, lrwork, liwork
+      integer, intent(out)      :: lwork, lrwork, liwork
       complex(8)                :: lwork_tmp
       real(8)                   :: lrwork_tmp
       integer(8)                :: N, info, minus_1!, me, nProcs
